requirements.txt
aiohttp==3.7.2
aiomysql==0.0.21
aioredis==1.3.1
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
discord.py==1.7.2
docopt==0.6.2
h11==0.11.0
h2==4.0.0
hiredis==2.0.0
hpack==4.0.0
hstspreload==2020.10.20
httpcore==0.12.0
httpx==0.16.1
hyperframe==6.0.0
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
loguru==0.5.3
monotonic==1.5
multidict==5.0.0
parso==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
PyMySQL==0.10.1 # upgrading causes errors
python-a2s==1.3.0
pytz==2021.1
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.15
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.25.11
wcwidth==0.2.5
websockets==8.1
yarl==1.6.2
python-decouple==3.4